removed the recursion and cut down on array creation. New version is ~5x faster and still passes tests:thumbsup:

```
# new version 100000 times
ok ~1.66 s (1 s + 661692338 ns)

# old version 100000 times
ok ~8.73 s (8 s + 731867943 ns)
```

on the other hand i think all the `Math.pow` calls and the ternary (for ignoring the middle "red" cell) make it look messy. i'll put some thought into how to make it look more elegant