@hughsk 2017-06-21
Wonderful, thanks @semibran! Merging/publishing — will add you to the repo/package on GitHub/npm too, feel free to make further optimisations/cleanup if you like :)